+++ packages/terra-modal-manager/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+* Added
+  * Added sizes `'fusion-small'`, `'fusion-medium'`, and `'fusion-large'`.
+
+## 6.64.0 - (July 19, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 6.63.0 - (June 5, 2023)
 
 * Changed